Time: 0 ms (100.00%) | Memory: 8.5 MB (67.76%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -4,8 +4,8 @@ public:
         if(x<0){
             return false;
         }
-        long long reversed =0;
-        long long temp=x;
+        long reversed =0;
+        long temp=x;
         while(temp!=0){
             int digit =temp%10;
             reversed=reversed*10+digit;
